+++ packages/core/solidity/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 0.5.4 (XXXX-XX-XX)
 
-- Add `account` and `paymaster` contract types for ERC-4337.
+- Add `account` and `paymaster` contract types for ERC-4337. ([#486](https://github.com/OpenZeppelin/contracts-wizard/pull/486))
+
+### Potentially breaking changes
+- Update pragma versions to 0.8.24. ([#486](https://github.com/OpenZeppelin/contracts-wizard/pull/486))
 
 ## 0.5.3 (2025-03-13)